Siesta 1.4.32 and 1.4.33, plus radfft removal from atom Changes: Removal of "use" statement for m_radfft from atom.f
Version: 1.4.33
Date: 2005/01/05 16:15 GMT
Title: Fixed bugs in diagk.F for writting of wave functions
Author: Pablo Ordejon <ordejon@icmab.es>

Some bugs introduced in the recent changes of diagk.F in the writting
of the wave functions are corrected.

Files:

diagk.F       All eigenvectors are computed when writing of wave 
              functions is required; the right information about
              eigenvalues is now passed.
writewave.F   Correct call to diagg.f to include new variable 
              neigwanted
version.F90 / CHANGES

=======================================================================
Version: 1.4.32
Date: 2004/12/20  8:30 GMT
Title: Fermi occupancy initialisation
Author: Julian Gale <julian@power.curtin.edu.au>

The occupancies are explicitly initialised to zero in fermid &
fermispin to ensure that the checking of occupancies in the build
of the density matrix is correct.

Files:

fermid.F      Initialisation of occupancies upto maxe added and
              cosmetic f90 changes made.
version.F90 / CHANGES