Coherent restarting for MD jobs (interim fix) The XV file contains the positions and velocities *at different times*. This poses problems for the continuation of MD runs.
Until we can re-design the MD output code, an interim fix has been
prepared:

For the velocity-Verlet algorithm (in routine 'verlet2'), the fix is to read the
'old' forces as well as the 'current' positions and 'old' velocities.
Forces are written to VERLET_FORCES at each step, and are read in the
first step if the XV file has been read. 

For the modified Verlet algorithm (in routine 'nose'), the fix is to read the
'old' positions, AND the values of the Nose variables (current and old).

The information is written to NOSE_RESTART at each step, and is read in the
first step if the XV file has been read. 

The code in ioxv.F has been put in a module, to hold the new variable
xv_file_read. 'dynamics.f' has also been turned into a module.

Fixes for the other MD flavors are forthcoming.