Fix matel to avoid non-zero gradients of overlaps in some cases Problem: (As restated by A. Postnikov in the Siesta list, after a message           by Hongjun Xiang <xhongjun@MAIL.USTC.EDU.CN>) --------
erroneous values of some gradients as delivered by matel.
I.e., the analytical gradient of the overlap of each basis function
with itself is not zero, contrary to expectation. An example for Al s-function:
  ia,io,ioa,ja,jo,joa:  1  1  1    1  1  1  xij=  0.0000  0.0000  0.0000
   S =  1.00000E+00  grad S =  0.00000E+00  0.00000E+00 -9.71745E-03
   T =  4.00390E-01  grad T =  0.00000E+00  0.00000E+00 -1.54155E-02
The values in the last column must be zero.

Jose Soler wrote:

The simplest and most apropriate solution to this problem is to simply
change the line
   DFFR0 = HUGE
 into
   DFFR0 = 0

That changes the boundary condition of the spline interpolation from
(d2FFR/dR2)_0=0 to (dFFR/dR)_0=0 and makes the error in the force of
order 1.e-12 rather than 1.e-2.

Done.