Bug fix in dlaed6 in dc_�lapack.f Javier Junquera got a new version of DLAED6 (part of LAPACK's divide-and-conquer package) from Ren-Cang Li <rcli@cs.uky.edu>. It partially solves some strange numerical instability problems. Codes produced by most compiler flag combinations now work (exceptions include  "-xP -mp1" for the Intel compiler).
It should be noted that the problem persists in many vendor-compiled
libraries (MKL, ACML, veclib, etc). It is not possible at this point
to use both the new fortran version and DLAED6 and the compiled
versions of the rest of the linear algebra routines. This will be
solved in a forthcoming patch.