New logic for k-point grid generation from kgrid_cutoff The previous algorithm created a diagonal supercell so that the length of the shortest vector was greater than twice the kgrid_cutoff.  But there might be an equivalent supercell (same volume, generating the same lattice points), but with shorter vectors (more "spherical").  Half of the shortest resulting lattice vector is the effective cutoff.  This effective cutoff was just reported, but the more spherical cell was not used anymore.  The kgrid was generated as the reciprocal cell of the chosen diagonal supercell.  In some cases, the program ended up using a much shorter cutoff than requested.
The fix, as suggested by Jose Soler, is to start the procedure with
the most "spherical" unit cell, thus generating an optimal supercell
in the spirit of the Moreno-Soler paper.

A number of other changes have been made to the code:

1. Minvec has been turned into a module, and the routine now reports
also the transformation matrix.  (Change to gchkmx needed, including
an auxiliary array to avoid aliasing of arguments).

2. Even though the program can find the optimal k-grid displacements,
those given by the user in a Monkhorst-Pack block are not overriden
(in case there are good reasons for the choice, such as the need to
have the gamma point in the set). The program will nevertheless print
a line telling the user about the more suitable displacement(s).

3. The program now reports the effective cutoff even for Gamma-point-only
calculations.

4. The multiplication factors of the effective supercell are made
explicitly positive (changing kdsc and mr after the call to idiag). This
might fix a long-standing issue with non-diagonal MP supercells.

Version info updated for development branch.

Note: There is *no provision for backward compatibility*, although it is
still possible to re-use the kgrid supercells reported by previous versions
to reproduce the old behavior for a given kgrid_cutoff.

To Do: Tidy up the kgridinit/kgrid code redundancy.

The manual has been updated to reflect the new algorithm outline, and
to stress the format of the MP supercell input.

Also, the manual now mentions the Divide and Conquer issue, and 
the version number on the cover has been changed to 2.0.X.