Fixes for Siesta-as-Server A number of fixes were needed for the proper functioning of the Siesta-as-Server package (also known as "SiestaAsSubroutine") (in addition to the purely cosmetic of openning the right CML context, already committed):
- iopipes.F90 had to be properly parallelized.

- fsiesta.f90 has been moved from Util?/... tothe main source directory
  in order to be compiled with the proper flushing routine.
  A couple of issues having to do with optional arguments have also been fixed.

- The makefile in Util/SiestaSubroutine/FmixMD/Src has been modified to
  deal with the new place for fsiesta.f90.

- Two new tests have been added. One of them tests the driving of a parallel
  Siesta calculation. A new script "test.sh" automates these tests.

In the main code, it has been noted that the "Forces" option (i.e.,
the computation of energy, forces, and stresses from geometry read
through pipes from a client program) should be assimilated to a
"variable cell" calculation, in the sense that the unit cell might be
different from call to call. Appropriate checks have been included to
regenerate k-point and supercell information if needed.

In the same vein, if the client inputs a null unit cell, Siesta should
generate an automatic one, at any step. A new routine "automatic_cell.f"
has been written and is now called at the beginning of the initialization
section, and after every reading of geometry from pipes. 

A null unit cell will be input whenever siesta_forces is not passed 
the optional "cell" parameter. This might happen during MD or relaxation
for a molecule. The actual unit cell used will then vary from step to
step. To avoid this, the cell should be passed explicitly in siesta_forces,
at every step (as it is done in the 'driver?' exampl).

Other changes:

- New Sys/intel9-cmkl8-mpi.make file for parallel compilation in Rocks cluster.

- A new script in Tests/Scripts automates the handling of mpd daemons
in an interactive single-nod run (when allowed).