Fix polarization basis specs for cases with two unocupied shells For atoms with a ground state configuration of the form
ns0 np0 (n-1)dX

a basis-spec of "szp" or "dzp" did not work, as the program tried to
polarize the s-shell, for which no PAOs were generated. Fixed by
explictly checking this special case in routine autobasis in basis_specs module.

(This case might apply to Pt, which in 2.0 is presumed to have a 
"6s0 5d10" configuration.)