Enhancements to the vibra package (Julian Gale)
An option has been introduced to compute infra-red intensities. In
order to do this the Born effective charges must have been computed
along with the force constants and the .BC file must be present in the
current directory. The intensities are estimated according to formula
(6) in the paper by Fernandez-Torre et al in J. Phys.  Chem. A, 108,
10535-10541 (2004).

Explicit copy of fdf.mod to deal with faulty arch.make's.

Removal of chkdim call from klines.

AG: Fixes for portability in Makefile and vibrator.f

Patches applied:

 * jgale@siesta.arch--2005/siesta-general--jgale--2.1--patch-1
   Call to chkdim removed from klines

 * jgale@siesta.arch--2005/siesta-general--jgale--2.1--patch-19
   Copying of fdf module to Vibra directory for build automated

 * jgale@siesta.arch--2005/siesta-general--jgale--2.1--patch-20
   Calculation of the IR intensities added to the Vibra package