Reparametrization of pseudopotentials * By changing the a and b parameters of the logarithmic grid, a new one with a more homogeneous overall grid-point separation can be used for the generation of basis sets and projectors. For example, by using a=5x10-4 and b=10, the grid point separations at r=0 and 10 bohrs are 0.005 and 0.01 bohrs, respectively. More points are needed to reach r's of the order of a hundred bohrs, but the extra computational effort is negligible.
This behavior is triggered by the boolean fdf symbol Reparametrize.Pseudos.

Further tuning can be achieved by the use of two other fdf symbols. The
processing syntax and defaults are:

         new_a = fdf_double("NewAParameter",5.0e-4_dp)
         new_b = fdf_double("NewBParameter",10.0_dp)

* New test: h2o_reparam

* Bug fix:

In routine findp in atom.f, the radius was wrongly computed as

r = b*(exp(a*(i-1)) + 1)

instead of

r = b*(exp(a*(i-1)) - 1)

The bug was dramatically exposed when using a re-parametrization in which
b is no longer small (10.0 vs 1.0e-4).