Scripting utilities Util/Scripting contains some proof-of-concept and prototype python modules and scripts.
Most of the functionality is based on the ASE package of the CAMPOS
project. See http://oldwww.fysik.dtu.dk/CAMPOS

The necessary modules are in the Siesta directory. The package can
also be installed in the appropriate Python directory using the
setup.py file provided. The scripts in the top directory could
then be transported anywhere.

Note that the location of the siesta executable must be specified
in the scripts at this point.

Directory scripts/ contains some general-purpose utilities.