Allow lifting of even/odd restrictions in radial grid The old code in arw.f needed to work with an odd number of points in the radial grid. That restriction has been lifted by coding a new "Simpson" integration routine (routine integrator in arw.f).
Those sections of atom.f in which the number of points is redefined
to make it odd (thus shifting the cutoff radii in the process) have
been wrapped in if tests. If the fdf symbol 'Restricted.Radial.Grid' is
set to .false., those sections will not be executed. 

Together with the 'Reparametrize.Pseudos' feature, this change will
allow a tighter mapping between the rc's specified externally (e.g.,
by a simplex driver) and those actually used by the program.

* New test:

    h2o_radialgrid