k-point sampling update when kgrid_cutoff is specified The user can specify a kgrid_cutoff or a Monkhorst-Pack grid to determine the k-point sampling. If there are unit-cell changes during the calculation, the k-point set does not change if a MP grid has been specified (the sampling accuracy changes implicitly).
On the other hand, the k-point set should change in principle if a
kgrid_cutoff has been given and the cell volume changes, but due to a
programming error the k-point grid was fixed after the first
iteration.

The error affects variable-cell calculations specifying a
kgrid_cutoff.  "Variable-cell" means that the unit cell changes
during a calculation, as in cell optimization and the MD flavors
Parrinello-Rahman, Nose-Parrinello-Rahman, and Annealing runs. 

A user might want to avoid sudden changes in the k-point set
while performing molecular dynamics runs or cell optimizations,
notably if the calculations are not very well converged. In that
case a MP grid, and not a kgrid_cutoff, should be specified.

A related problem that has been fixed in this patch is the 
possible overriding of the user's MP displacement specification after
the first iteration.

There is no backwards compatibility after this bugfix. To recover the
old behavior when a kgrid_cutoff was specified, the user should check
the old output, extract the information about the MP grid generated
automatically by the program, and put it in an MP block.

(Regeneration of reference outputs)