Created struct_init for initial geometry setup. New test force_2 Struct_init has been detached from siesta_init to perform a self-contained geometry setup.
Notes:

* Ns is now set in initatom, and it is "intent(in)" in redata.

* Those variables which have to do with structural input 
are read twice: once in struct_init, and then in redata, pending a
re-structuring of the latter. In some cases local variables are
used.

* iza is still set up in initatomlists. This means that some of the
structural output has to be kept in siesta_init for now. The fix is to
compute iza in coor or in read_struct.

* va and vcell are zeroed out if an XV file is not used. But the
zmatrix i/o routine is in principle able to read vcell (?).

* Put the calls to struct_init and initatomlists closer.
Do the array initialization inside of initatomlists.

New test "force_2" for force-constant calculation with a smaller
runtime than "force_constants"

(Note: the initial coordinate output and other details are now output
in a different order from that in the base version)