Consolidate geometry updates at the end of loop * Move comment about start of geometry loop to the right place in  siesta.F
* For idyn cases 6 (force constant), 7 (phonon interface), and 8
(server calculation), move the coordinate updates from the beginning
to the end of the geometry loop. The logic of geometry output to file XV
has been maintained. For cases 6 and 7, any output to a MD history file
is now done with the current (distorted) coordinates, instead of with
the restored (undisplaced) coordinates (bug of original version).

Note that the logic of XV output for the true MD cases has been maintained.
(next step's geometry, current step's velocities).

The MD output (together with any pixmol output) has been encapsulated
in a new file write_md_record.F