New m_energies and m_steps modules. Back to old k-point behavior More steps in the slimming of siesta_todo:
* Module m_energies contains the energy components
* Module m_steps contains inicoor, fincoor, istp, final
* Variable dDmax has been turned into an argument of
  mixer and scfconvergence_test

*The old code had a bug in the update of k-points with cell changes:
the kscell matrix was kept constant, instead of re-calculating it.
This was fixed in the re-structured version, but of course the fix
leads to differences in the outputs of some tests, notably var_cell.

The bug has been re-activated in kpoint_grid.F90 by commenting out
the update of kscell.

The resetting of kscell to the unit matrix in case of gamma
calculations has been deferred for the same reason.